// EmployeeReferralManagement/app/models/candidate.ts
export class Candidate {
    id: number;
    name: string;
    email: string;
    phone: string;
    alternateNumber: string;
    experience: string;
    role: string;
    referredBy: string;
    candidateId: string;
    status: string;
    resume: string;
    isChecked: boolean;
    checkedBy: string;
}